+++ lib/assert.js
@@ -25,15 +25,12 @@ const {
   isDeepEqual,
   isDeepStrictEqual
 } = require('internal/util/comparisons');
-const {
-  AssertionError,
-  errorCache,
-  codes: {
-    ERR_AMBIGUOUS_ARGUMENT,
-    ERR_INVALID_ARG_TYPE,
-    ERR_INVALID_RETURN_VALUE
-  }
-} = require('internal/errors');
+const { codes: {
+  ERR_AMBIGUOUS_ARGUMENT,
+  ERR_INVALID_ARG_TYPE,
+  ERR_INVALID_RETURN_VALUE
+} } = require('internal/errors');
+const { AssertionError, errorCache } = require('internal/assert');
 const { openSync, closeSync, readSync } = require('fs');
 const { inspect, types: { isPromise, isRegExp } } = require('util');
 const { EOL } = require('os');
