+++ test/parallel/test-assert.js
@@ -29,7 +29,7 @@ const common = require('../common');
 const assert = require('assert');
 const { EOL } = require('os');
 const EventEmitter = require('events');
-const { errorCache } = require('internal/errors');
+const { errorCache } = require('internal/assert');
 const { writeFileSync, unlinkSync } = require('fs');
 const { inspect } = require('util');
 const a = assert;
